requirements.txt
absl-py==2.0.0
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
efficientnet==1.0.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
idna==3.4
image-classifiers==1.0.0
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
keras==2.14.0
Keras-Applications==1.0.8
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.8
networkx==3.2
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
segmentation-models==1.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
